feat(book): extract StyledInput, PasswordField, BottomDrawer, BottomController into injected-ui
Starts the popups category: these 4 were the only popups components
already wired into the book, via a deep relative import straight into
packages/mask/popups/components (bypassing injected-ui entirely,
unlike every other book demo). All four were already fully pure - no
plugin-infra, no live hooks, no @masknet/shared, no lingui macros - so
this is a straight move plus updating their ~17 production call sites
(popups modals and pages) to import from @masknet/injected-ui/* instead.

BottomDrawer was re-exported through packages/mask/popups/components/
index.ts; that export is dropped and its 6 consumers now import it
directly from injected-ui, consistent with the rest of the codebase.

No new dependencies (packages/mask already depends on injected-ui).

Book demos and their descriptions updated to reference the new source
location.

Follow-up to #12445-#12453, moving from the injection category (now
fully extracted) into popups.
