+++ apps/book/src/demos/popups/BottomController.demo.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react'
 import { Button, Stack, Typography } from '@mui/material'
-import { BottomController } from '../../../../../packages/mask/popups/components/BottomController/index.js'
+import { BottomController } from '@masknet/injected-ui/BottomController'
 
 export const meta = {
     title: 'BottomController',
     description:
-        'Fixed action bar docked to the bottom of a popups page (packages/mask/popups/components/BottomController). Note: it renders `position: fixed` to the viewport — in the real extension that viewport is the ~400px popup window, so it docks to the bottom of this page here instead.',
+        'Fixed action bar docked to the bottom of a popups page (packages/injected-ui/src/BottomController.tsx). Note: it renders `position: fixed` to the viewport — in the real extension that viewport is the ~400px popup window, so it docks to the bottom of this page here instead.',
 }
 
 export default function BottomControllerDemo() {
