+++ packages/mask/popups/components/index.ts
@@ -1,5 +1,4 @@
 export * from './ActionModal/index.js'
 export * from './NormalHeader/index.js'
-export * from './BottomDrawer/index.js'
 export * from './WalletBalance/index.js'
 export * from './TokenPicker/index.js'
