+++ packages/mask/popups/modals/AddContactModal/index.tsx
@@ -9,7 +9,7 @@ import { isValidAddress } from '@masknet/web3-shared-evm'
 import { IconButton, InputAdornment, Typography, useTheme } from '@mui/material'
 import { evm } from '@masknet/web3-providers'
 import { useContacts } from '@masknet/web3-hooks-base'
-import { BottomDrawer, type BottomDrawerProps } from '../../components/index.js'
+import { BottomDrawer, type BottomDrawerProps } from '@masknet/injected-ui/BottomDrawer'
 import { isSameAddress } from '@masknet/web3-shared-base'
 import { Icons } from '@masknet/icons'
 import { Trans, useLingui } from '@lingui/react/macro'
