+++ packages/mask/popups/modals/ChangeBackupPasswordModal/index.tsx
@@ -2,7 +2,7 @@ import { memo, useCallback } from 'react'
 import { ActionModal, type ActionModalBaseProps } from '../../components/index.js'
 import { UserContext } from '../../../shared-ui/index.js'
 import { Box, Typography, useTheme } from '@mui/material'
-import { PasswordField } from '../../components/PasswordField/index.js'
+import { PasswordField } from '@masknet/injected-ui/PasswordField'
 import { Controller, useForm } from 'react-hook-form'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
