+++ packages/mask/popups/pages/Personas/ExportPrivateKey/index.tsx
@@ -5,7 +5,7 @@ import { Box, Button, Link, Typography, useTheme } from '@mui/material'
 import { PersonaContext } from '@masknet/shared'
 import Services from '#services'
 import { useAsync, useCopyToClipboard } from 'react-use'
-import { BottomController } from '../../../components/BottomController/index.js'
+import { BottomController } from '@masknet/injected-ui/BottomController'
 import { useNavigate, Link as RouterLink } from 'react-router-dom'
 import { PopupRoutes } from '@masknet/shared-base'
 import { ActionButton, useSnackbar } from '@masknet/theme'
