+++ packages/mask/popups/pages/Personas/Logout/index.tsx
@@ -12,8 +12,8 @@ import { memo, useCallback, useMemo, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { useAsyncFn } from 'react-use'
 import { UserContext } from '../../../../shared-ui/index.js'
-import { BottomController } from '../../../components/BottomController/index.js'
-import { PasswordField } from '../../../components/PasswordField/index.js'
+import { BottomController } from '@masknet/injected-ui/BottomController'
+import { PasswordField } from '@masknet/injected-ui/PasswordField'
 import { PersonaAvatar } from '../../../components/PersonaAvatar/index.js'
 import { useTitle } from '../../../hooks/index.js'
