+++ packages/mask/popups/pages/Personas/PersonaAvatarSetting/index.tsx
@@ -6,7 +6,7 @@ import { Box, Button, Slider, Typography } from '@mui/material'
 import { Icons } from '@masknet/icons'
 import { ActionButton, makeStyles, useSnackbar } from '@masknet/theme'
 import { PersonaContext } from '@masknet/shared'
-import { BottomController } from '../../../components/BottomController/index.js'
+import { BottomController } from '@masknet/injected-ui/BottomController'
 import { NormalHeader } from '../../../components/index.js'
 import { PopupRoutes } from '@masknet/shared-base'
 import { useTitle } from '../../../hooks/index.js'
